Refactor QueryParamsGenerator from Generator
