+++ lib/murker/generator.rb
@@ -1,10 +1,13 @@
-require 'json-schema-generator'
 require 'murker/path_params_generator'
+require 'murker/query_params_generator'
+require 'murker/utils'
 
 module Murker
   # Generates OpenAPI v3 schema for Interaction
   # Output is Ruby object representing schema, it may be serialized to yml|json
   class Generator
+    include Murker::Utils
+
     OPEN_API_VERSION = '3.0.0'.freeze
     OPEN_API_INFO = {
       'title' => 'Generated by Murker',
@@ -76,18 +79,7 @@ module Murker
     end
 
     def query_params_schema
-      return nil unless query_params
-
-      query_params.map do |name, value|
-        {
-          'in' => 'query',
-          'name' => name,
-          'description' => name,
-          'schema' => schema_by_query_param_value(value),
-          'required' => true,
-          'example' => value,
-        }
-      end
+      QueryParamsGenerator.call(query_params)
     end
 
     def response_schema
@@ -101,22 +93,6 @@ module Murker
           },
         },
       }
-    end
-
-    def schema_by_query_param_value(value)
-      # simple value - string
-      # array
-      # hash
-    end
-
-    def schema_by_object(obj)
-      serialized_schema = JSON::SchemaGenerator.generate(
-        '', # comment for generated json
-        JSON.dump(obj),
-        schema_version: 'draft4',
-      )
-      schema_object = JSON.parse(serialized_schema)
-      remove_unexpected_fields(schema_object)
     end
 
     def remove_unexpected_fields(schema_object)
